pkg: introduce the config package
Move the config logic, currently handling the configuration flags,
into a proper package. Benefits of this move:
1. better testability
2. enable code reuse
3. enable future refactorings/changes
4. open the opportunity for code simplification
5. allows us to take the chance to rewrite the unit tests, shedding a
   dep - we should really use the simple testing package here unless
   something more powerful is clearly needed.
